+++ Odontotal/Odontotal/src/main/java/JuniorsDH/Odontotal/Domain/Paciente.java
@@ -2,7 +2,6 @@ package JuniorsDH.Odontotal.Domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
 import java.time.LocalDate;
@@ -29,13 +28,13 @@ public class Paciente extends Usuario {
     @CreationTimestamp
     private Date fechaCreacion;
 
-    public Paciente(Long id, String email, String password, String nombre, String apellido, String documento, LocalDate fechaNacimiento, Genero genero, int telefono, String urlImagen, Domicilio domicilio, UsuarioRol rol, Long id1, Boolean validado, Date fechaCreacion) {
+    public Paciente(Long id, String email, String password, String nombre, String apellido, String documento, LocalDate fechaNacimiento, Genero genero, Long telefono, String urlImagen, Domicilio domicilio, UsuarioRol rol, Long id1, Boolean validado, Date fechaCreacion) {
         super(id, email, password, nombre, apellido, documento, fechaNacimiento, genero, telefono, urlImagen, domicilio, rol);
         this.validado = validado;
         this.fechaCreacion = fechaCreacion;
     }
 
-    public Paciente(String email, String password, String nombre, String apellido, String documento, LocalDate fechaNacimiento, Genero genero, int telefono, String urlImagen, Domicilio domicilio, UsuarioRol rol, Boolean validado, Date fechaCreacion) {
+    public Paciente(String email, String password, String nombre, String apellido, String documento, LocalDate fechaNacimiento, Genero genero, Long telefono, String urlImagen, Domicilio domicilio, UsuarioRol rol, Boolean validado, Date fechaCreacion) {
         super(email, password, nombre, apellido, documento, fechaNacimiento, genero, telefono, urlImagen, domicilio, rol);
         this.validado = validado;
         this.fechaCreacion = fechaCreacion;
