+++ Odontotal/Odontotal/src/main/java/JuniorsDH/Odontotal/Dto/OdontologoDto.java
@@ -1,8 +1,5 @@
 package JuniorsDH.Odontotal.Dto;
 
-import JuniorsDH.Odontotal.Domain.Domicilio;
-import JuniorsDH.Odontotal.Domain.UsuarioRol;
-
 import java.time.LocalDate;
 
 public class OdontologoDto extends UsuarioDto {
@@ -12,13 +9,13 @@ public class OdontologoDto extends UsuarioDto {
 
     private String urlImagen;
 
-    public OdontologoDto(Long id, String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, int telefono, String calle, String numero, String localidad, String provincia, String rol, String matricula, String especialidad, String urlImagen) {
+    public OdontologoDto(Long id, String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, Long telefono, String calle, String numero, String localidad, String provincia, String rol, String matricula, String especialidad, String urlImagen) {
         super(id, nombre, apellido, email, password, documento, fechaNacimiento, genero, telefono, calle, numero, localidad, provincia, rol, urlImagen);
         this.matricula = matricula;
         this.especialidad = especialidad;
     }
 
-    public OdontologoDto(String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, int telefono, String calle, String numero, String localidad, String provincia, String rol, String matricula, String especialidad, String urlImagen) {
+    public OdontologoDto(String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, Long telefono, String calle, String numero, String localidad, String provincia, String rol, String matricula, String especialidad, String urlImagen) {
         super(nombre, apellido, email, password, documento, fechaNacimiento, genero, telefono, calle, numero, localidad, provincia, rol, urlImagen);
         this.matricula = matricula;
         this.especialidad = especialidad;
