+++ Odontotal/Odontotal/src/main/java/JuniorsDH/Odontotal/Dto/PacienteDto.java
@@ -1,9 +1,5 @@
 package JuniorsDH.Odontotal.Dto;
 
-import JuniorsDH.Odontotal.Domain.Domicilio;
-import JuniorsDH.Odontotal.Domain.UsuarioRol;
-
-import javax.persistence.Column;
 import java.time.LocalDate;
 import java.util.Date;
 
@@ -13,13 +9,13 @@ public class PacienteDto extends UsuarioDto{
 
     private Date fechaCreacion;
 
-    public PacienteDto(Long id, String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, int telefono, String calle, String numero, String localidad, String provincia, String rol, String urlImagen, Boolean validado, Date fechaCreacion) {
+    public PacienteDto(Long id, String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, Long telefono, String calle, String numero, String localidad, String provincia, String rol, String urlImagen, Boolean validado, Date fechaCreacion) {
         super(id, nombre, apellido, email, password, documento, fechaNacimiento, genero, telefono, calle, numero, localidad, provincia, rol, urlImagen);
         this.validado = validado;
         this.fechaCreacion = fechaCreacion;
     }
 
-    public PacienteDto(String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, int telefono, String calle, String numero, String localidad, String provincia, String rol, String urlImagen, Boolean validado, Date fechaCreacion) {
+    public PacienteDto(String nombre, String apellido, String email, String password, String documento, LocalDate fechaNacimiento, String genero, Long telefono, String calle, String numero, String localidad, String provincia, String rol, String urlImagen, Boolean validado, Date fechaCreacion) {
         super(nombre, apellido, email, password, documento, fechaNacimiento, genero, telefono, calle, numero, localidad, provincia, rol, urlImagen);
         this.validado = validado;
         this.fechaCreacion = fechaCreacion;
